added cui banner component
