+++ src/app-components/cui-banner.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const CuiBanner = () => (
+  <div className='cui-banner row no-gutters'>
+    <b>Highest Possible Classification is CUI</b>
+  </div>
+);
+
+export default CuiBanner;
