--- src/common/footer/footer.scss
@@ -1,16 +1,0 @@
-@import '../../style/variables';
-
-.footer {
-  background-color: $color-primary-dark;
-  bottom: 0;
-  color: $color-white;
-  min-height: 40px;
-  padding-top: 9px;
-  position: fixed;
-  width: 100%;
-  z-index: 10000;
-
-  p {
-    margin-bottom: 0;
-  }
-}
