+++ src/css/index.scss
@@ -120,3 +120,14 @@ input[disabled] {
 .ag-theme-balham .ag-cell-inline-editing {
   height: 100% !important;
 }
+
+.cui-banner {
+  background-color: $color-black;
+  color: $color-white;
+  justify-content: center !important;
+  height: 20px;
+
+  .cui-banner-dropdown {
+    z-index: 8888;
+  }
+}
